usersContext works and fixed issue where index was refreshing
