+++ src/App.js
@@ -2,18 +2,16 @@
 import { Route, Switch } from 'react-router-dom';
 import IndexPage from './Containers/IndexPage';
 import UserDetail from './Components/UserDetail';
+import { UsersProvider } from './UsersContext';
 
 function App() {
 	return (
 		<div>
 			<Switch>
-				{/* {location.state !== undefined ? ( */}
-				<Route path='/users/:id' component={UserDetail} />
-				{/* ) : (
-					<Redirect to='/' />
-				)} */}
-
-				<Route path='/' component={IndexPage} />
+				<UsersProvider>
+					<Route path='/users/:id' component={UserDetail} />
+					<Route path='/' component={IndexPage} />
+				</UsersProvider>
 			</Switch>
 		</div>
 	);
